sweet alert works

+++ client/src/views/Boards.vue
@@ -26,6 +26,8 @@
 </template>
 
 <script>
+import Swal from "sweetalert2";
+
 export default {
   name: "boards",
   mounted() {
@@ -53,7 +55,23 @@ export default {
       this.form = false;
     },
     deleteBoard(id) {
-      this.$store.dispatch("deleteBoard", id);
+      Swal.fire({
+        title: "Are you sure you want to delete this board?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+        backdrop: `
+      rgba(123, 0, 16, 0.4)
+  `
+      }).then(result => {
+        if (result.value) {
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+          this.$store.dispatch("deleteBoard", id);
+        }
+      });
     }
   }
 };
